+++ src/libs/etherlib/parameter.cpp
@@ -135,7 +135,7 @@ void CParameter::registerClass(void) {
     if (been_here) return;
     been_here = true;
 
-    uint32_t fieldNum = 1000;
+    size_t fieldNum = 1000;
     ADD_FIELD(CParameter, "schema",  T_NUMBER, ++fieldNum);
     ADD_FIELD(CParameter, "deleted", T_BOOL,  ++fieldNum);
     ADD_FIELD(CParameter, "showing", T_BOOL,  ++fieldNum);
