+++ src/libs/test-libs/etherlib/convert/options.cpp
@@ -16,7 +16,7 @@
 CParams params[] = {
     CParams("",        "This is what the program does.\n"),
 };
-uint32_t nParams = sizeof(params) / sizeof(CParams);
+size_t nParams = sizeof(params) / sizeof(CParams);
 
 //---------------------------------------------------------------------------------------------------
 bool COptions::parseArguments(string_q& command) {
