+++ src/libs/test-libs/utillib/colorTest/options.cpp
@@ -16,7 +16,7 @@
 CParams params[] = {
     CParams("",             "Simple test program.\n"),
 };
-uint32_t nParams = sizeof(params) / sizeof(CParams);
+size_t nParams = sizeof(params) / sizeof(CParams);
 
 //---------------------------------------------------------------------------------------------------
 bool COptions::parseArguments(string_q& command) {
