+++ src/libs/utillib/database.h
@@ -70,7 +70,7 @@ namespace qblocks {
         // forces implementation
         virtual string_q getType(void) const = 0;
 
-        bool Lock(const string_q& fn, const string_q& mode, uint32_t obeyLock);
+        bool Lock(const string_q& fn, const string_q& mode, size_t obeyLock);
         bool ReLock(const string_q& mode);
         void Release(void);
         void Close(void);
@@ -147,8 +147,7 @@ namespace qblocks {
     //----------------------------------------------------------------------
     extern size_t stringToAsciiFile(const string_q& fileName, const string_q& contents);
     extern string_q binaryFileToString(const string_q& filename);
-    extern bool asciiFileToBuffer(const string_q& filename, size_t& nChars, string_q *buffer,
-                                  uint32_t maxLines = INT_MAX);
+    extern bool asciiFileToBuffer(const string_q& filename, size_t& nChars, string_q *buffer, size_t maxLines = INT_MAX);
     extern uint64_t appendToAsciiFile(const string_q& fileName, const string_q& addContents);
 
     //----------------------------------------------------------------------
@@ -169,7 +168,7 @@ namespace qblocks {
     extern void writeTheCode(const string_q& fileName, const string_q& code, const string_q& ns = "", bool spaces = true);
 
     //----------------------------------------------------------------------
-    inline bool asciiFileToBuffer(const string_q& filename, string_q *contents, uint32_t maxLines = INT_MAX) {
+    inline bool asciiFileToBuffer(const string_q& filename, string_q *contents, size_t maxLines = INT_MAX) {
         size_t unused = 0;
         return asciiFileToBuffer(filename, unused, contents, maxLines);
     }
