+++ src/libs/utillib/options_base.h
@@ -106,7 +106,7 @@ namespace qblocks {
 
     //--------------------------------------------------------------------------------
     extern CParams *paramsPtr;
-    extern uint32_t& nParamsRef;
+    extern size_t& nParamsRef;
     extern COptionsBase *pOptions;
 
     extern bool isEnabled(uint32_t q);
